requirements.txt
Flask
gunicorn
Pillow
PyPDF2
exifread
python-magic
